split off near-dupe test data generators
